MAP-1450 Add training environment deployment (#71)
* MAP-1450 Add training environment deployment

This commit introduces a new preprod training environment to the CircleCI configuration and adds corresponding Helm values for deployment. The changes include updates to `.circleci/config.yml` to integrate the new environment and a new `values-train.yaml` file to specify environment-specific configurations.

* MAP-1450 Remove unnecessary Jira environment type

The `jira_env_type` field was removed from the `deploy_preprod` job configuration in the CircleCI config file. This streamlines the configuration by eliminating a redundant parameter.

* MAP-1450 Rename deploy_preprod to deploy_train

Updated the deployment step in the CircleCI configuration from 'deploy_preprod' to 'deploy_train'. This adjustment aligns the naming conventions and targets the 'train' environment for the deployment process.