fix: Remove indefinite caching in locations service
We thought that the service was spawned once pre request, but it is
actually a long-lived object so this code would cause the constant
values to be kept in memory until pod restart, so any new values
would never be retrieved. Instead we will remove this code and let
Redis handle the caching.

MAP-1489